[feat] 게시물의 상세 내용 확인 성공
